requirements.txt
torch>=1.10.0
torchvision>=0.11.0
torchaudio>=0.10.0
numpy>=1.21.0
matplotlib>=3.4.0
Pillow>=8.3.0
scikit-image>=0.18.0
